Centralise ScalePresentationArea() in ScreenManager. Fix scaling on mobile devices.
