+++ Templates/MonoGame.Templates.CSharp/content/MonoGame.2D.StartKit/___SafeGameName___.Core/Screens/AboutScreen.cs
@@ -1,18 +1,4 @@
-#region File Description
-//-----------------------------------------------------------------------------
-// OptionsMenuScreen.cs
-//
-// Microsoft XNA Community Game Platform
-// Copyright (C) Microsoft Corporation. All rights reserved.
-//-----------------------------------------------------------------------------
-#endregion
-
-#region Using Statements
-using ___SafeGameName___.Core;
 using ___SafeGameName___.Core.Localization;
-using Microsoft.Xna.Framework;
-using System;
-#endregion
 
 namespace ___SafeGameName___.Screens;
 
@@ -23,15 +9,8 @@ namespace ___SafeGameName___.Screens;
 /// </summary>
 class AboutScreen : MenuScreen
 {
-    #region Fields
-
     MenuEntry builtWithMonoGameMenuEntry;
     MenuEntry monoGameWebsiteMenuEntry;
-
-    #endregion
-
-    #region Initialization
-
 
     /// <summary>
     /// Constructor.
@@ -42,7 +21,6 @@ class AboutScreen : MenuScreen
         // Create our menu entries.
         builtWithMonoGameMenuEntry = new MenuEntry("#BuiltWithMonoGame", false);
         monoGameWebsiteMenuEntry = new MenuEntry(Resources.MonoGameSite);
-        // Need to add attribution link <a href="https://www.flaticon.com/free-icons/backpack" title="backpack icons">Backpack icons created by Freepik - Flaticon</a>
         MenuEntry back = new MenuEntry(Resources.Back);
 
         // Hook up menu event handlers.
@@ -54,10 +32,6 @@ class AboutScreen : MenuScreen
         MenuEntries.Add(monoGameWebsiteMenuEntry);
         MenuEntries.Add(back);
     }
-
-    #endregion
-
-    #region Handle Input
 
     /// <summary>
     /// Event handler for when the MonoGame Website menu entry is selected.
@@ -76,6 +50,4 @@ class AboutScreen : MenuScreen
         // More tweaks required to make this work on all platforms :)
         System.Diagnostics.Process.Start(new System.Diagnostics.ProcessStartInfo(url) { UseShellExecute = true });
     }
-
-    #endregion
 }
