+++ Templates/MonoGame.Templates.CSharp/content/MonoGame.2D.StartKit/___SafeGameName___.Core/Screens/MessageBoxScreen.cs
@@ -171,11 +171,9 @@ class MessageBoxScreen : GameScreen
         // Darken down any other screens that were drawn beneath the popup.
         ScreenManager.FadeBackBufferToBlack(TransitionAlpha * 2 / 3);
 
-        // Center the message text in the viewport.
-        Viewport viewport = ScreenManager.GraphicsDevice.Viewport;
-        Vector2 viewportSize = new Vector2(viewport.Width, viewport.Height);
+        // Center the message text in the BaseScreenSize. The GlobalTransformation will scale everything for us.
         Vector2 textSize = font.MeasureString(message);
-        Vector2 textPosition = (viewportSize - textSize) / 2;
+        Vector2 textPosition = (ScreenManager.BaseScreenSize - textSize) / 2;
 
         // The background includes a border somewhat larger than the text itself.
         const int hPad = 32;
@@ -189,7 +187,7 @@ class MessageBoxScreen : GameScreen
         // Fade the popup alpha during transitions.
         Color color = Color.White * TransitionAlpha;
 
-        spriteBatch.Begin();
+        spriteBatch.Begin(SpriteSortMode.Deferred, null, null, null, null, null, ScreenManager.GlobalTransformation);
 
         // Draw the background rectangle.
         spriteBatch.Draw(gradientTexture, backgroundRectangle, color);
